requirements.txt
fastapi
uvicorn[standard]
httpx
requests
bme680
transformers
llama-cpp-python
huggingface-hub
